+++ src/core/configmanager.h
@@ -28,12 +28,14 @@ class ConfigManager : public QObject
                          QObject *parent = nullptr);
   void save();
   void readGeneralSettings();
-  QFont getEditorFontFamily() const;
+  QString getEditorFontFamily() const;
   void setEditorFontFamily(const QString &fontname);
   QString getStyle() const;
   void setStyle(const QString &stylename);
   int getEditorFontSize() const;
-  void setEditorFontSize(const int fontsize);
+  void setEditorFontSize(const int &fontsize);
+  QString getColorTheme() const;
+  void setColorTheme(const QString &ctname);
 
  private:
   QSettings *settings;
@@ -41,6 +43,7 @@ class ConfigManager : public QObject
   QString editorFontFamily;
   QString style;
   int editorFontSize;
+  QString colorTheme;
 };
 
 #endif  // CONFIGMANAGER_H